add responsive design

+++ src/Components/Auth/Login/Login.js
@@ -1,20 +1,54 @@
 import React, { useState } from 'react'
-import { Box, Typography, TextField, Button, CircularProgress} from '@mui/material'
+import { Box, Typography, TextField, Button, CircularProgress, InputAdornment} from '@mui/material'
+import { RecaptchaVerifier, signInWithPhoneNumber } from 'firebase/auth'
+import { auth } from '../../../firebase'
 import { useNavigate } from 'react-router-dom'
-//import { Link } from 'react-router-dom'
 
 function Login() {
     const navigate = useNavigate()
     const [phone, setPhone] = useState('')
+    const [otp, setOtp] = useState('')
     const [loading, setLoading] = useState(false)
+    const [otpSent, setOtpSent] = useState(false)
 
     const handleLogin = () => {
         setLoading(true)
-        setTimeout(() => {
+        window.recaptchaVerifier = new RecaptchaVerifier(auth, 'recaptcha-container', {
+            size : 'invisible',
+            callback : () => {
+                const mobPhone = `+91 ${phone}`
+                const appVerifier = window.recaptchaVerifier
+                signInWithPhoneNumber(auth, mobPhone, appVerifier)
+                .then((data)=> {
+                    setLoading(false)
+                    setOtpSent(true)
+                    window.confirmationResult = data;
+                })
+                .catch(error=> {
+                    setLoading(false)
+                    console.log('error', error)
+                })
+            },
+        })
+        window.recaptchaVerifier.verify();
+    }
+
+    const handleVerifyOTP = () => {
+        setLoading(true)
+        window.confirmationResult.confirm(otp)
+        .then(()=> {
             setLoading(false)
             localStorage.setItem('loggedIn', true)
+            localStorage.setItem('mobile', phone)
             navigate('/')
-        }, 2000);
+            setOtp('')
+            setPhone('')
+            setOtpSent(false)
+        })
+        .catch(error=> {
+            setLoading(false)
+            console.log('error', error)
+        })
     }
 
     return (
@@ -35,33 +69,38 @@ function Login() {
                 variant="outlined"
                 label="Phone"
                 size="small"
+                type="number"
                 value={phone}
+                disabled={otpSent}
                 onChange={(e)=> setPhone(e.target.value)}
                 placeholder='Enter mobile number'
+                InputProps={{
+                    startAdornment: <InputAdornment position="start">+91</InputAdornment>,
+                  }}
                 sx={{
                     mb : '30px',
-                    // '& fieldset' : {
-                    // borderRadius : '30px',
-                    // p : '10px 10px'
-                    // }
                 }}
             />
-            {/* <TextField 
-                fullWidth
-                variant="outlined"
-                label="Password"
-                size="small"
-                placeholder='Enter Password'
-                sx={{
-                    mb : '30px',
-                    // '& fieldset' : {
-                    // borderRadius : '30px',
-                    // }
-                }}
-            /> */}
+            {
+                otpSent && (
+                    <TextField 
+                        fullWidth
+                        value={otp}
+                        onChange={(e)=> setOtp(e.target.value)}
+                        type="number"
+                        variant="outlined"
+                        label="OTP"
+                        size="small"
+                        placeholder='Enter Password'
+                        sx={{
+                            mb : '30px',
+                        }}
+                    /> 
+                )
+            }
             <Button
                 variant="contained"
-                onClick={handleLogin}
+                onClick={()=> otpSent ? handleVerifyOTP() :  handleLogin()}
                 disabled={loading}
             >
                 {
@@ -71,24 +110,14 @@ function Login() {
                         size={18}
                     />
                     :
-                    "Proceed"
+                    otpSent
+                    ?
+                    "Verify OTP"
+                    :
+                    "Generate OTP"
                 }
             </Button>
-            <Box
-                display="flex"
-                alignItems="center"
-                justifyContent="center"
-                sx={{
-                    mt : '10px'
-                }}
-            >
-                <Typography variant="p">New user, </Typography>
-                {/* <Link
-                    to='/signup'
-                >
-                    Signup
-                </Link> */}
-            </Box>
+            <div id="recaptcha-container"></div>
         </Box>
     )
 }
